Fixed segmentation fault on whitespace input.
Before, inputting whitespace (any number of spaces) on it's own would cause a
segfault, (e.g.)

`user@linux: $ ____`

(where the '_'s are spaces) would cause a segfault in the `cd_handler`
function due to argv[0] potentially being null.

I added an `ltrim` function, which trims whitespace beginning from the
beginning of the char* passed. I used this to make sure a command isn't
parsed if it is just whitespace.